src/test/java/com/collaboration/config/EmailServiceImplTest.java
package com.collaboration.config;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.doThrow;
import static org.mockito.Mockito.never;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentCaptor;
import org.mockito.Captor;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.mail.javamail.JavaMailSender;
import org.springframework.mail.javamail.MimeMessageHelper;
import org.springframework.mail.javamail.MimeMessagePreparator;

import jakarta.mail.MessagingException;
import jakarta.mail.internet.MimeMessage;

@ExtendWith(MockitoExtension.class)
public class EmailServiceImplTest {

    @Mock
    private JavaMailSender mailSender;

    @Mock
    private MimeMessage mimeMessage;

    @Mock
    private MimeMessageHelper mimeMessageHelper;

    @Captor
    private ArgumentCaptor<MimeMessagePreparator> mimeMessagePreparatorCaptor;

    @InjectMocks
    private EmailServiceImpl emailService;

    @Disabled("Funktioniert leider nicht")
    @Test
    public void testSendSimpleMessageSuccess() throws Exception {
        // Vorbereitung
        when(mailSender.createMimeMessage()).thenReturn(mimeMessage);
        when(new MimeMessageHelper(mimeMessage, true)).thenReturn(mimeMessageHelper);

        // Testen
        emailService.sendSimpleMessage("test@example.com", "Test Subject", "Test Text");

        // Überprüfen
        verify(mailSender).send(mimeMessagePreparatorCaptor.capture());
        MimeMessagePreparator preparator = mimeMessagePreparatorCaptor.getValue();
        preparator.prepare(mimeMessage);

        verify(mimeMessage).setSubject("Test Subject");
        verify(mimeMessageHelper).setFrom("admin@example.com");
        verify(mimeMessageHelper).setTo("test@example.com");
        verify(mimeMessageHelper).setText("Test Text", true);
    }

    @Disabled("Funktioniert leider nicht")
    @Test
    public void testSendSimpleMessageFailure() throws Exception {
        // Vorbereitung
        when(mailSender.createMimeMessage()).thenReturn(mimeMessage);
        when(new MimeMessageHelper(mimeMessage, true)).thenReturn(mimeMessageHelper);
        doThrow(new MessagingException("Test Exception")).when(mimeMessageHelper).setText(any(), any());

        // Testen
        emailService.sendSimpleMessage("test@example.com", "Test Subject", "Test Text");

        // Überprüfen
        verify(mailSender, never()).send(any(MimeMessagePreparator.class));
    }

    @Disabled("Funktioniert leider nicht")
    @Test
    public void testInitialization() {
        // Überprüfen, ob die Klasse korrekt initialisiert wurde
        assertNotNull(emailService.getMailSender());
        assertEquals("admin@example.com", emailService.getAdminEMail());
    }
}
